put float together except for frac part

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: vice-wra <vice-wra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/18 13:28:20 by jblue-da          #+#    #+#             */
-/*   Updated: 2019/04/16 14:46:28 by vice-wra         ###   ########.fr       */
+/*   Updated: 2019/04/16 17:11:15 by vice-wra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@
 
 
 union {
-		long double d_num;
+		double d_num;
 		long long ll_num;
 	} t;
